upload

+++ build_env.sh
@@ -1,4 +1,6 @@
 #!/bin/sh
 
-cp backend/.env.example backend/.env
+if [ ! -f backend/.env ]; then
+  cp backend/.env.example backend/.env
+fi
 docker-compose build && docker-compose up -d
